Finalizada 90%

+++ flutter_tracktrail_app/lib/domain/entities/user_database_entity.dart
@@ -8,7 +8,7 @@ class UserDatabaseEntity {
   final DateTime dateOfBirth;
   final String sex;
   final double height;
-  final int role;
+  final double role;
   final String? avatar;
 
   UserDatabaseEntity({
